Add in a predesigned level
Make it possible to specify a level when constructing a new GemGrid.
Also add a Gem constructor to choose shape and colour, otherwise the
previous is a little pointless.

We should consider the time investment into making a level editor. I
think they are pretty important for a puzzle game…